# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Binder Project Planner API
  version: 0.1.0
  description: REST API for the local Binder Project Planner application.
servers:
  - url: http://localhost:3001
paths:
...
  /binders:
    get:
      operationId: listBinders
      summary: List all binders
      description: >-
        Returns the complete initial binder-summary collection (no
        pagination), sorted by updatedAt descending and then by binder UUID
        ascending for deterministic ordering.
      responses:
        '200':
          description: The binder-summary collection in its documented sort order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BinderSummary'
    post:
      operationId: createBinder
      summary: Create a new binder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBinderRequest'
      responses:
        '201':
          description: The binder was created.
          headers:
            Location:
              description: The path of the newly created binder resource.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Binder'
        '400':
          description: The request body did not match the documented schema.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: A binder with the same case-insensitively normalized name already exists.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /binders/{binderId}:
    parameters:
      - $ref: '#/components/parameters/binderId'
    get:
      operationId: getBinder
      summary: Get one binder's details
      description: >-
        Returns binder details for the shared binder context (story 7) and
        the Edit Details tab.
      responses:
        '200':
          description: The binder details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Binder'
        '400':
          description: The binderId path parameter is not a well-formed UUID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: No binder exists with the given id.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    patch:
      operationId: updateBinder
      summary: Update a binder's details
      description: >-
        Applies a partial update to the documented dirty fields (story 7
        covers name/width/height/pages; later stories add notes, preview
        page, dimension, and lock-state fields).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBinderRequest'
      responses:
        '200':
          description: The complete persisted binder after applying the update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Binder'
        '400':
          description: >-
            The binderId path parameter is not a well-formed UUID, or the
            request body did not match the documented schema.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: No binder exists with the given id.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: A binder with the same case-insensitively normalized name already exists.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /binders/{binderId}/cards:
    parameters:
      - $ref: '#/components/parameters/binderId'
    get:
      operationId: listBinderCards
      summary: List a binder's cards
      description: >-
        Returns every binder-owned card, placed and unplaced, without image
        bytes (story 11).
      responses:
        '200':
          description: The binder's cards.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        '400':
          description: The binderId path parameter is not a well-formed UUID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: No binder exists with the given id.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      operationId: createBinderCard
      summary: Assign a TCGdex card to a binder slot
      description: >-
        Creates a binder-owned card from a normalized TCGdex catalog result
        and its target placement (story 11). The backend assigns the
        `tcgdex` source, downloads and installs the card's image, and
        validates placement against the binder's current dimensions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: The card was created and placed.
          headers:
            Location:
              description: The path of the newly created card resource.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          description: >-
            The binderId path parameter is not a well-formed UUID, or the
            request body did not match the documented schema or a valid
            placement within the binder's current dimensions.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: No binder exists with the given id.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: The target placement is already occupied by another card.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '502':
          description: The upstream TCGdex image download failed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '504':
          description: The upstream TCGdex image download timed out.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /binders/{binderId}/art:
    parameters:
      - $ref: '#/components/parameters/binderId'
    get:
      operationId: listBinderArt
      summary: List a binder's multi-slot art
      description: >-
        Returns every binder-owned multi-slot-art record, placed and
        unplaced, without image bytes. Art creation does not exist yet
        (story 25), so this always returns an empty array today.
      responses:
        '200':
          description: The binder's multi-slot art.
          content:
            application/json:
              schema:
                type: array
                items: {}
        '400':
          description: The binderId path parameter is not a well-formed UUID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: No binder exists with the given id.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  parameters:
    binderId:
      name: binderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
  schemas:
...
    ProblemDetails:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
    CreateBinderRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - width
        - height
        - pages
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Trimmed on the backend; case-insensitive uniqueness is enforced there.
        width:
          type: integer
          minimum: 1
          description: Number of slot columns per binder side.
        height:
          type: integer
          minimum: 1
          description: Number of slot rows per binder side.
        pages:
          type: integer
          minimum: 1
          description: Stored binder-page count. Physical pages range from 1 through 2 * pages.
    UpdateBinderRequest:
      type: object
      additionalProperties: false
      description: >-
        A partial update; only supplied fields are changed. Story 7 covers
        name/width/height/pages; later stories (20-24) add notes, preview
        page, dimension, and lock-state fields to this same request schema.
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Trimmed on the backend; case-insensitive uniqueness is enforced there.
        width:
          type: integer
          minimum: 1
          description: Number of slot columns per binder side.
        height:
          type: integer
          minimum: 1
          description: Number of slot rows per binder side.
        pages:
          type: integer
          minimum: 1
          description: Stored binder-page count. Physical pages range from 1 through 2 * pages.
    Binder:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - width
        - height
        - pages
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        pages:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BinderSummary:
      type: object
      additionalProperties: false
      description: >-
        The lightweight binder representation used by the home-page binder
        list (story 5). Later stories (20, 22) extend this schema with only
        the additional preview and completion-metric data those features
        require, rather than the complete card and multi-slot-art graph.
      required:
        - id
        - name
        - width
        - height
        - pages
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        pages:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlacementCoordinates:
      type: object
      additionalProperties: false
      description: >-
        One-based physical page, row, and column. All three are populated
        for a placed item and all three are null for an unplaced item
        (story 11's slot assignment always supplies all three).
      required:
        - physicalPage
        - row
        - column
      properties:
        physicalPage:
          type: integer
          minimum: 1
          nullable: true
        row:
          type: integer
          minimum: 1
          nullable: true
        column:
          type: integer
          minimum: 1
          nullable: true
...
    CreateCardRequest:
      type: object
      additionalProperties: false
      description: >-
        Creates a TCGdex-sourced binder card (story 11). The backend does
        not refetch card details before saving; it assigns `source`
        server-side and downloads the image from `imageUrl` itself.
      required:
        - name
        - setName
        - localNumber
        - providerCardId
        - providerSetId
        - imageUrl
        - placement
      properties:
        name:
          type: string
        setName:
          type: string
          nullable: true
        localNumber:
          type: string
          nullable: true
        providerCardId:
          type: string
        providerSetId:
          type: string
        imageUrl:
          type: string
          format: uri
        variation:
          type: string
          maxLength: 50
          nullable: true
        placement:
          $ref: '#/components/schemas/PlacementCoordinates'
    Card:
      type: object
      additionalProperties: false
      description: The complete persisted representation of a binder-owned card.
      required:
        - id
        - binderId
        - name
        - setName
        - localNumber
        - source
        - providerCardId
        - providerSetId
        - variation
        - placement
        - imageUrl
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        binderId:
          type: string
          format: uuid
        name:
          type: string
        setName:
          type: string
          nullable: true
        localNumber:
          type: string
          nullable: true
        source:
          type: string
          enum:
            - tcgdex
            - custom
        providerCardId:
          type: string
          nullable: true
        providerSetId:
          type: string
          nullable: true
        variation:
          type: string
          nullable: true
        placement:
          $ref: '#/components/schemas/PlacementCoordinates'
        imageUrl:
          type: string
          format: uri-reference
          description: The backend's own `/cards/{cardId}/image` endpoint URL.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time